requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
numpy>=1.24.0
pytz>=2023.3
polygon-api-client>=1.1.1
jsonschema>=4.17.0
pyyaml>=6.0
pyarrow>=12.0.0